Fix floating-point confidence comparison and build config
- Change DetectionResult.confidence from float to double to avoid float→double
  precision loss causing 0.7f < 0.70 comparison to fail
- Raise Signal 1 contribution from 0.5 to 0.70 so frozen-body+moving-look
  alone triggers detection without requiring additional signals
- Fix root build.gradle.kts plugin declaration syntax
- Temporarily exclude fabric/neoforge from settings.gradle.kts (require
  separate build toolchains; re-add when building locally)
- All 15 common module unit tests pass

+++ anti-freecam/common/src/main/java/dev/thewindows/antifreecam/common/detection/DetectionResult.java
@@ -1,12 +1,12 @@
 package dev.thewindows.antifreecam.common.detection;
 
-public record DetectionResult(boolean detected, float confidence, String reason) {
+public record DetectionResult(boolean detected, double confidence, String reason) {
 
     public static DetectionResult notDetected() {
-        return new DetectionResult(false, 0f, "No freecam signals");
+        return new DetectionResult(false, 0.0, "No freecam signals");
     }
 
     public static DetectionResult insufficient() {
-        return new DetectionResult(false, 0f, "Insufficient data");
+        return new DetectionResult(false, 0.0, "Insufficient data");
     }
 }
